revert: Reverted PR from wide format due to error in exporting process. Keep the delimiter function to dynamically adapt the Kinexon file splitting.
This reverts commit 2a86e87e3cd3c61ca572aded8102d33df1d9a6ed.

The previous PR that changed the format of the Kinexon file has been reverted due to an error in the exporting process. However, the delimiter function that was added in the PR is still useful for dynamically adapting the file splitting. This commit reverts the file format changes and keeps the delimiter function in place.