feat(kinexon-io): Accelerate CSV reading with league ID and delimiter specification
This commit introduces enhancements to the Kinexon IO module for faster reading of CSV files. Two small features have been added: the ability to use a league ID as an identifier and the option to specify the delimiter for reading CSV files.

Additionally, there is a breaking change related to the functions get_meta_data and read_position_data_csv. Instead of line-based reading, these functions now utilize vectorized functions provided by the pandas library, resulting in a significant speedup of approximately x8.